Addressing Jacks and Navneets comments
